add separate env for e2e
* Adds separate env for e2e that will enable us to have different mock
record, schema and configuration so that we can change them according
to what we want to e2e test without messing up dev app. Closes #366

* Also allows us to change dev mock data and configration without
breaking e2e tests unnecessarily

* Uses `ng e2e` that serves the example app builtin which removes the
`sleep` and occasional failures on travis.
