+++ example/configs/index.ts
@@ -1,0 +1,4 @@
+// ALWAYS access config through `environment`, DO NOT import from here
+
+export { defaultConfig } from './config';
+export { e2eConfig } from './config.e2e';
