+++ example/environments/environment.e2e.ts
@@ -1,0 +1,7 @@
+import { e2eConfig } from '../configs';
+
+export const environment = {
+  production: true,
+  mockDataFolder: 'mock-data-e2e',
+  editorConfig: e2eConfig
+};
