+++ example/environments/environment.prod.ts
@@ -1,3 +1,8 @@
+import { defaultConfig } from '../configs';
+
 export const environment = {
-  production: true
+  production: true,
+  editorConfig: defaultConfig,
+  mockDataFolder: 'mock-data'
 };
+
