+++ example/environments/environment.ts
@@ -3,6 +3,10 @@
 // `ng build --env=prod` then `environment.prod.ts` will be used instead.
 // The list of which env maps to which file can be found in `angular-cli.json`.
 
+import { defaultConfig } from '../configs';
+
 export const environment = {
-  production: false
+  production: false,
+  editorConfig: defaultConfig,
+  mockDataFolder: 'mock-data'
 };
