+++ package.json
@@ -18,14 +18,12 @@
   "scripts": {
     "build": "gulp build",
     "start": "npm run build && npm run serve",
-    "serve": "ng serve",
+    "serve:": "ng serve",  
     "lint": "tslint --project ./src/tsconfig.json \"src/**/*.ts\"",
     "test:unit": "ng test --code-coverage --sourcemaps=false --watch=false",
-    "test": "npm-run-all build test:unit e2e:travis",
+    "test": "npm-run-all build test:unit e2e",
     "pree2e": "webdriver-manager update",
-    "e2e": "protractor",
-    "e2e:travis": "npm-run-all --parallel -r serve e2e:travis:delay",
-    "e2e:travis:delay": "sleep 15 && npm run e2e",
+    "e2e": "ng e2e --env=e2e",
     "preversion": "npm test",
     "version": "npm run build",
     "copy": "copyfiles LICENSE README.md package.json dist",
